+++ front/lib/deprecated_action_configurations_server_only.ts
@@ -1,31 +1,9 @@
-import type {
-  AgentActionConfigurationType,
-  AgentConfigurationType,
-  TemplateAgentConfigurationType,
-} from "@dust-tt/types";
+import type { AgentActionConfigurationType } from "@dust-tt/types";
 import { assertNever } from "@dust-tt/types";
 
 import type { AgentConfigurationWithoutActionsType } from "@app/lib/api/assistant/configuration";
 import { AgentConfiguration } from "@app/lib/models";
 import logger from "@app/logger/logger";
-
-export function deprecatedGetFirstActionConfiguration(
-  config: AgentConfigurationType | TemplateAgentConfigurationType
-) {
-  if (config.actions.length > 1) {
-    logger.warn(
-      {
-        agentConfigurationId: "sId" in config ? config.sId : "template",
-        agentConfigurationName: config.name,
-      },
-      "Multiple actions are not supported yet. The first action will be used."
-    );
-  }
-  if (config.actions.length) {
-    return config.actions[0];
-  }
-  return null;
-}
 
 // TODO(@fontanierh) Temporary, to remove.
 // This is a shadow write while we invert the relationship between configuration and actions.
