+++ front/pages/api/w/[wId]/assistant/agent_configurations/index.ts
@@ -17,7 +17,6 @@ import type * as t from "io-ts";
 import * as reporter from "io-ts-reporters";
 import type { NextApiRequest, NextApiResponse } from "next";
 
-import { deprecatedMaybeShadowWriteFirstActionOnAgentConfiguration } from "@app/lib/action_configurations";
 import { trackAssistantCreated } from "@app/lib/amplitude/node";
 import { getAgentUsage } from "@app/lib/api/assistant/agent_usage";
 import {
@@ -29,6 +28,7 @@ import {
 } from "@app/lib/api/assistant/configuration";
 import { getAgentsRecentAuthors } from "@app/lib/api/assistant/recent_authors";
 import { Authenticator, getSession } from "@app/lib/auth";
+import { deprecatedMaybeShadowWriteFirstActionOnAgentConfiguration } from "@app/lib/deprecated_action_configurations_server_only";
 import { safeRedisClient } from "@app/lib/redis";
 import { apiError, withLogging } from "@app/logger/withlogging";
